% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{correlaciones}
\alias{correlaciones}
\title{Calcula las correlaciones}
\usage{
correlaciones(ds, plot = FALSE)
}
\arguments{
\item{ds}{Objeto de tipo Dataset}

\item{plot}{Logical. Fijado en TRUE, mostrará las correlaciones en un gráfico}
}
\value{
Matriz de correlaciones
}
\description{
Calcula las correlaciones entre pares de variables (columnas) de un \code{\linkS4class{Dataset}}.
Se seleccionarán las columnas númericas automáticamente
}
\seealso{
\code{infmutuas}
}
