% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetManager.R
\docType{class}
\name{Dataset-class}
\alias{Dataset-class}
\title{Clase Dataset}
\description{
Clase para guardar dataframes con un nombre
}
\details{
Este paquete inclute viñetas con ejemplos de uso de este paquete
}
\section{Slots}{

\describe{
\item{\code{data}}{Dataframe que contenga los datos}

\item{\code{name}}{Nombre asignado al Dataset}
}}

