% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetManager.R
\name{dataset}
\alias{dataset}
\title{Creación de Datasets}
\usage{
dataset(datos, id, factorizar = 0)
}
\arguments{
\item{datos}{Un vector, matriz o dataframe de datos. Mínimo dos elementos
en el vector y dos filas en la matriz o dataframe}

\item{id}{(Opcional) Nombre que se le proporcionará al Dataset (string).
En caso de no declararlo, se usará un número (el cuál incrementará por cada
Dataset creado sin nombre)}

\item{factorizar}{Las columnas que contengas menos valores diferentes que
este parámetro serán factorizadas}
}
\value{
Objeto \code{\linkS4class{Dataset}} creado, con nombre y factorizado (si procede)
}
\description{
Función para crear \code{\linkS4class{Dataset}} de este paquete y poder operar con ellas
}
\examples{
dataset(c(3,3,5,8,9,5), id="prueba", factorizar = 5)
}
\seealso{
\code{\link{leer_datos}} para crear Dataset desde un fichero de datos
}
