% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifications.R
\name{estandarizar}
\alias{estandarizar}
\title{Estandarización de variables}
\usage{
estandarizar(ds, columnas)
}
\arguments{
\item{ds}{Objeto de tipo Dataset.}

\item{columnas}{El índice de columnas a las que se aplicará la estandarización.
Las columnas deberán ser numéricas y no deberán estar ya factorizadas.}
}
\value{
Un nuevo \code{\linkS4class{Dataset}} con las columnas proporcionadas
estandarizadas
}
\description{
Estandariza las columnas señaladas de un \code{\linkS4class{Dataset}}.
Los valores seguirán una distribución normal con media 0 y desviación estandar 1.
}
\seealso{
\code{\link{normalizar}}
}
