% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifications.R
\name{filtrar}
\alias{filtrar}
\title{Filtrar elementos de un Dataset}
\usage{
filtrar(ds, columna, funcion)
}
\arguments{
\item{ds}{Objeto de tipo Dataset.}

\item{columna}{El índice de columna al que se aplicará la función de filtrado.}

\item{funcion}{Función de filtrado. Deberá constar de un único argumento de entrada
que será un vector, y su salida (return) deberá ser un vector de TRUE o FALSE si se
cumple el criterio de la función por cada elemento del vector de entrada.}
}
\value{
Un nuevo \code{\linkS4class{Dataset}} con las filas filtradas.
}
\description{
Selecciona las filas de un \code{\linkS4class{Dataset}} que cumplen un criterio en una columna.
Este criterio se especificará a través de una función
}
\examples{
\dontrun{
filtrar(ds, 3, function(x) x == "Radio")
filtrar(ds, 3, function(x) x == "Radio" || x == "Television")
filtrar(ds, 1, function(x) x > 4.7)
filtrar(ds, 4, function(x) x < 3)
filtrar(ds, 2, function(x) x == TRUE)
}
}
