% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{graficoBoxplot}
\alias{graficoBoxplot}
\title{Mostrar gráficamente Boxplot}
\usage{
graficoBoxplot(ds)
}
\arguments{
\item{ds}{Objeto de tipo Dataset}
}
\description{
Se visualizará mediante un gráfico los boxplot correspondientes a cada columna.
Se aplicará solo a las columnas numéricas, las no numéricas se omitiran
automáticamente en el gráfico.
}
