% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetManager.R
\name{guardar_datos}
\alias{guardar_datos}
\title{Guardar los datos}
\usage{
guardar_datos(path, ds)
}
\arguments{
\item{path}{Nombre del fichero con el que se desea guardar la información}

\item{ds}{Objeto de tipo \code{\linkS4class{Dataset}}}
}
\description{
Guarda los datos que tiene el \code{\linkS4class{Dataset}} en formato csv
}
