% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{infmutuas}
\alias{infmutuas}
\title{Calcula las informaciones mutuas}
\usage{
infmutuas(ds, logical = FALSE)
}
\arguments{
\item{ds}{Objeto de tipo Dataset}

\item{logical}{Variable para controlar si se tendrán en cuenta también
las columnas del Dataset que sean de tipo logical. Por defecto, FALSE (no
se tendrán en cuenta). Si se fija en TRUE, se considerarán.}
}
\value{
Matriz de correlaciones
}
\description{
Calcula las informaciones mutuas entre pares de variables (columnas) de un \code{\linkS4class{Dataset}}.
Se seleccionarán las columnas de valores numéricos discretos y factores
automáticamente.
}
