% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetManager.R
\name{leer_datos}
\alias{leer_datos}
\title{Cargar los datos}
\usage{
leer_datos(path, id, encabezado = FALSE, sep = ",", dec = ".", factorizar = 0)
}
\arguments{
\item{path}{Ruta del fichero csv (o derivados) a leer}

\item{id}{(Opcional) Nombre que se le proporcionará al Dataset (string).
En caso de no declararlo, se usará un número (el cuál incrementará por cada
Dataset creado sin nombre)}

\item{encabezado}{Logical. Indicar si el fichero contiene en la primera línea
los nombres de las columnas. Por defecto es \code{FALSE}}

\item{sep}{Caracter que separa las columnas en el fichero. Por defecto, ','}

\item{dec}{Caracter que indica como están representados los decimales.
Por defecto, '.'}

\item{factorizar}{Las columnas que contengas menos valores diferentes que
este parámetro serán factorizadas. Por defecto no se factorizará (0)}
}
\value{
Objecto de tipo \code{\linkS4class{Dataset}} con los datos del fichero
}
\description{
Lee los datos de un fichero csv y crea objeto \code{\linkS4class{Dataset}}
}
\details{
La función está preparada para leer ficheros csv por defecto. Sin embargo,
se puede leer tsv y otros formatos indicando el separador adecuado y caracter
para números decimales
}
\examples{
\dontrun{
leer_datos(mi_path, encabezado = TRUE)
}
}
\seealso{
\code{\link{dataset}} para crear objetos Dataset desde variables
}
