% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetManager.R
\name{nombres_columna}
\alias{nombres_columna}
\title{Cambiar los nombres de las columnas}
\usage{
nombres_columna(ds, nombres, columnas = NULL)
}
\arguments{
\item{ds}{Objeto de tipo \code{\linkS4class{Dataset}}}

\item{nombres}{Un vector con los nombres de las columnas. Ha de contener un nombre
por cada columna existente en el Dataset si no se define el parámetro \code{"columnas"}.
En ese caso el vector podrá contener menos nombres.}

\item{columnas}{Por defecto NULL. Será definido si se desea reemplazar el nombre de
una o varias columnas del dataframe. Para ello se pasará al argumento un número o
vector de números los cuales serán los índices de las columnas deseadas. En ellas
se asignarán los nombres indicados en el parámetro \code{"nombres"}}
}
\description{
Esta función cambiará los nombres de las columnas de un \code{\linkS4class{Dataset}}.
Pueden cambiarse los nombres de todas las columnas a la vez o de las seleccionadas únicamente
}
\examples{
\dontrun{
nombres_columna(ds, c("1", "2", "3", "4"))
nombres_columna(ds, c("5","6"), c(1,2))
nombres_columna(ds3, 9, 3)
}
}
