% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifications.R
\name{normalizar}
\alias{normalizar}
\title{Normalización de variables}
\usage{
normalizar(ds, columnas)
}
\arguments{
\item{ds}{Objeto de tipo Dataset.}

\item{columnas}{El índice de columnas a las que se aplicará la normalización.
Las columnas deberán ser numéricas y no deberán estar ya factorizadas.}
}
\value{
Un nuevo \code{\linkS4class{Dataset}} con las columnas proporcionadas
normalizadas
}
\description{
Normaliza las columnas señaladas de un \code{\linkS4class{Dataset}}.
Los valores estarán en el rango de entre 0 y 1 incluídos.
}
\seealso{
\code{\link{estandarizar}}
}
