% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetManager.R
\docType{methods}
\name{print,Dataset-method}
\alias{print,Dataset-method}
\title{Imprimir Dataset}
\usage{
\S4method{print}{Dataset}(x)
}
\arguments{
\item{x}{Objeto Dataset a imprimir}
}
\description{
Imprime el \code{\linkS4class{Dataset}}, junto a su nombre y número de filas/columnas
}
