% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smer.R
\docType{package}
\name{smer}
\alias{smer}
\title{smer: Easy-to-use dataframe statistical analysis}
\description{
El objetivo de este paquete es facilitar la aplicación de cálculos y estadísticos en datos a usuarios que dispongan de poca experiencia en el ámbito de la programación. También es adecuado para conseguir los resultados deseados automáticamente en pocas líneas de código sin necesidad de programar nada.
Entre las funciones disponibles, se encuentran la lectura y escritura de datasets, cálculo de la varianza, curva ROC, discretización, normalización y estandarización de variables, cálculo de la correlación, cálculo de la entropía y distintos gráficos que representen los resultados.
}
\section{Funciones del paquete smer}{

Gestión de Dataset: \code{\link{dataset}}, \code{\link{leer_datos}},
\code{\link{guardar_datos}}, \code{\link{nombres_columna}},
\code{\link{print}}
}

