% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{varianzas}
\alias{varianzas}
\title{Varianza del Dataset}
\usage{
varianzas(ds)
}
\arguments{
\item{ds}{Objeto de tipo Dataset.}
}
\value{
Vector con la varianza correspondiente a cada columna.
Si la columna de datos no es numérica, devolverá NA en su respectiva posición.
}
\description{
Calcula la varianza de cada columna de un \code{\linkS4class{Dataset}}.
}
